PedPartitionType and PedPartitionFlag are enums, so treat them as longs.
This fixes the same stack alignment problems as we were having with PedUnit.

+++ include/pydisk.h
@@ -141,7 +141,7 @@ typedef struct {
     PyObject *disk;                /* _ped.Disk */
     PyObject *geom;                /* _ped.Geometry */
     int num;
-    long long type;                /* PedPartitionType */
+    long type;                     /* PedPartitionType */
     PyObject *fs_type;             /* _ped.FileSystemType */
 } _ped_Partition;
